Template script for adding metadata to R datafiles
A short script using functions from the package 'dataMeta' to create a data dictionary for an R data frame and add it to the data frame attributes. May be useful particularly for the covariates data frame (since it probably wont change a whole lot) and up-dated field data merges in order to easily keep track up of which data sets are included. The template uses the Herbarium dataset I put together from CPNWH.