Create extrac_quadpolyID_function.py
I adapted your python code into a function with an additional if statement which creates a new geodatabase if the output path does not lead to one. I used to have an additional statement for iterating geodatabases containing field data, but it got a little unweildly for my python skills... always an option down the road. I figured most new field data would be a shapefile anyway...  This script probably needs work, Im noticing redundancies and issues with the else statement as I load this...