Create get_extent_shp
Function for easily creating an extent shapefile from field data input. (a set_extent function is in the works which will have the ability to call internal system files from the package of shapefiles for IDFG regions, IDFG GMUs, and 100k eCognition polygons which could be used as iterators to easily restrict analysis).